feat : 로또 번호 출력
기능 목록
 - Lotto 번호 출력
 - 오름차순 정렬

Lotto class에 numbers List를 오름 차순 정렬을 위해 ArrayList로 초기화 하고 print하는 메서드를 추가하였다.
